Rewrote some prime generating functions to be u64 instead of generic

+++ src/problem/highly_divisible_triangle_number.rs
@@ -1,7 +1,7 @@
 use super::Problem;
 use crate::{number_theory::linear_sum, primes::prime_factors};
 
-fn num_factors(n: u32) -> u32 {
+fn num_factors(n: u64) -> u64 {
     let prime_factors = prime_factors(n);
     prime_factors.into_iter().map(|(_, n)| n + 1).product()
 }
@@ -18,7 +18,7 @@ impl Problem for HighlyDivisibleTriangleNumber {
     }
 
     fn solve(&self) -> String {
-        const TARGET_NUM_DIVISORS: u32 = 500;
+        const TARGET_NUM_DIVISORS: u64 = 500;
 
         let mut n = 7;
         while num_factors(linear_sum(n)) <= TARGET_NUM_DIVISORS {
